refactor: make FrameAllocator methods allocation free
